requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
convertdate==2.2.0
DateTime==4.3
flake8==3.8.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
holidays==0.10.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
livereload==2.6.1
lunr==0.5.6
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1
more-itertools==8.2.0
mysql-connector==2.2.9
mysqlclient==1.4.6
nltk==3.4.5
nr.collections==0.0.1
nr.databind==0.0.4
nr.interface==0.0.1
nr.metaclass==0.0.4
nr.parsing.date==0.0.1
nr.pylang.utils==0.0.1
nr.stream==0.0.2
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pathtools==0.1.2
pluggy==0.13.1
progressbar2==3.51.4
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyMeeus==0.3.7
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-utils==2.4.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
tornado==6.0.4
typing==3.7.4.1
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.9
Werkzeug==1.0.0
yapf==0.29.0
zipp==3.1.0
zope.interface==5.0.1
